BED-4937 moved Users and associated files to shared ui
