+++ packages/javascript/bh-shared-ui/rollup.config.js
@@ -66,5 +66,6 @@ export default {
         'lodash/cloneDeep',
         'lodash/isEmpty',
         'lodash/startCase',
+        'lodash/find',
     ],
 };
