+++ packages/javascript/bh-shared-ui/src/components/CreateUserDialog/CreateUserDialog.tsx
@@ -16,7 +16,7 @@
 
 import { Dialog, DialogTitle } from '@mui/material';
 import React from 'react';
-import CreateUserForm from './CreateUserForm';
+import CreateUserForm from '../CreateUserForm';
 import { CreateUserRequest } from "js-client-library";
 
 const CreateUserDialog: React.FC<{
