+++ packages/javascript/bh-shared-ui/src/components/CreateUserDialog/index.ts
@@ -1,0 +1,3 @@
+export { default } from './CreateUserDialog';
+
+export * from './CreateUserDialog'
