+++ packages/javascript/bh-shared-ui/src/components/CreateUserForm/index.ts
@@ -1,0 +1,3 @@
+export { default } from './CreateUserForm';
+
+export * from './CreateUserForm';
