+++ packages/javascript/bh-shared-ui/src/components/UserActionsMenu/index.ts
@@ -1,0 +1,3 @@
+export { default } from './UserActionsMenu';
+
+export * from './UserActionsMenu';
