+++ packages/javascript/bh-shared-ui/src/views/Users/index.ts
@@ -1,0 +1,3 @@
+export { default } from './Users';
+
+export * from './Users';
