+++ packages/javascript/bh-shared-ui/src/views/index.ts
@@ -20,8 +20,12 @@ export { default as UserProfile } from './UserProfile';
 
 export { default as SSOConfiguration } from './SSOConfiguration';
 
+export { default as Users } from './Users';
+
 export * from './Explore';
 
 export * from './DataQuality';
 
 export * from './Explore/ExploreSearch';
+
+export * from './Users'
